refactor of systemtime plugin
